Added documentation groups.
Added edge constructors with implicit and explicit vertex initialization to make it easier to write new edge types.

Injected 5D and 6D matrix and vertex types into the Eigen namespace.

Added nicer antialiased line rasterizer.

Added proper support for constant vertices.

Reduced memory usage in covariances calculation.

Added Eigenvalue calculation capabilities.

Added a tool fur calculating exact numbers of FLOPs in sparse matrix operations.

Added new BA examples.

Optimized includes for faster builds (using Deheader).

Added scripts folder, moved Doxyfile over to there.

Added some new block matrix predicates, namely CUberBlockMatrix::b_UpperTriangular(), CUberBlockMatrix::b_UpperBlockTriangular(), CUberBlockMatrix::b_StrictlyUpperBlockTriangular(), CUberBlockMatrix::b_LowerTriangular(), CUberBlockMatrix::b_LowerBlockTriangular(), CUberBlockMatrix::b_StrictlyLowerBlockTriangular(), CUberBlockMatrix::b_No_Diagonal_Blocks(), CUberBlockMatrix::b_All_Diagonal_Blocks(), CUberBlockMatrix::b_SymmetricBlockStructure() and CUberBlockMatrix::b_SymmetricBlocksOnly().

Added an implementation of CUberBlockMatrix::TriangularViewOf() which makes a shallow (optionally deep) copy of matrix triangles. Supports upper or lower triangle, including the diagonal.

Renamed CUberBlockMatrix::p_Convert_to_Sparse_UpperDiagonal_Debug() to the correct name p_Convert_to_Sparse_UpperTriangular_Debug(). Backwards compatibility was not retained in this case as it is a debugging function, presumably not used in production code.

Added native implementation of A+A^T for AMD ordering on the block matrices, via CUberBlockMatrix::p_BlockStructure_SumWithSelfTransposeNoDiag_to_Sparse32(), CUberBlockMatrix::p_BlockStructure_SumWithSelfTransposeNoDiag_to_Sparse(), CUberBlockMatrix::BlockStructure_SumWithSelfTranspose().

Added CUberBlockMatrix::Get_Diagonal() which reads a diagonal of a block matrix.

Added CUberBlockMatrix::InverseOf_Symmteric() which inverts symmetric block matrix (using dense code, suitable for block diagonal matrices).

Added CUberBlockMatrix::UpperTriangularTranspose_Solve_FBS() which enables solving with sparse right hand side.

Added CUberBlockMatrix::ExtendTopLeftTo() for putting empty blocks in the top left corner of the matrix.

Improved MatrixMarket I/O support.

Added fbs_ut::CWrap3 and fbs_ut::CDummyAlgSpecializer for writing simple algorithms specialized for fixed block sizes.

Added conversions from CUberBlockMatrix to VBR and BSR BLAS formats.

+++ src/cxsparse/cs_sqr.c
@@ -1,4 +1,4 @@
-#include "cs.h"
+#include "cxsparse/cs.h"
 /* compute nnz(V) = S->lnz, S->pinv, S->leftmost, S->m2 from A and S->parent */
 static CS_INT cs_vcount (const cs *A, css *S)
 {
@@ -81,7 +81,7 @@ css *cs_sqr (CS_INT order, const cs *A, CS_INT qr)
     }
     else
     {
-        S->unz = 4*(A->p [n]) + n ;         /* for LU factorization only, */
+        S->unz = (double)(4*(A->p [n]) + n) ;         /* for LU factorization only, */ // swine - fixed a watning in x64
         S->lnz = S->unz ;                   /* guess nnz(L) and nnz(U) */
     }
     return (ok ? S : cs_sfree (S)) ;        /* return result S */
